backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2==2.9.11
groq==1.0.0
requests==2.31.0
asgiref==3.7.2
sqlparse==0.4.4
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
httpx==0.27.0
pydantic==2.6.0
anyio==4.3.0
cryptography==42.0.5